temperature sensor fault state bugfix
temperature sensor was set to can bus as 8 bit integer, but it was declare as 16 bit. now this is fixed.
Oil cooler is also disabled if oil temperature sensor is faulty

+++ sw/inc/sensor.h
@@ -26,7 +26,7 @@ typedef uint8_t sensor_state_e;
 /// @file: Implements a module for reading analog sensor data
 typedef struct {
 	uv_moving_aver_st avg;
-	int16_t value;
+	int8_t value;
 	bool hyst_warning_enabled;
 	bool hyst_error_enabled;
 	sensor_state_e state;
@@ -36,23 +36,23 @@ typedef struct {
 	esb_emcy_e emcy_error;
 	esb_emcy_e emcy_fault;
 	uv_adc_channels_e adc;
-	bool (*get_data)(uv_adc_channels_e chn, int16_t *dest);
+	bool (*get_data)(uv_adc_channels_e chn, int8_t *dest);
 } sensor_st;
 
 
 
 void sensor_init(sensor_st *this, uv_adc_channels_e adc_chn, uint16_t avg_count,
-		esb_emcy_e emcy_fault, bool (*get_data_ptr)(uv_adc_channels_e chn, int16_t *dest));
+		esb_emcy_e emcy_fault, bool (*get_data_ptr)(uv_adc_channels_e chn, int8_t *dest));
 
-void sensor_set_warning(sensor_st *this, int16_t trigger_value, int16_t hysteresis,
+void sensor_set_warning(sensor_st *this, int8_t trigger_value, int8_t hysteresis,
 		bool invert, esb_emcy_e emcy_warning);
 
-void sensor_set_error(sensor_st *this, int16_t trigger_value, int16_t hysteresis,
+void sensor_set_error(sensor_st *this, int8_t trigger_value, int8_t hysteresis,
 		bool invert, esb_emcy_e emcy_error);
 
 void sensor_step(sensor_st *this, uint16_t step_ms);
 
-static inline int16_t sensor_get_value(const sensor_st *this) {
+static inline int8_t sensor_get_value(const sensor_st *this) {
 	return this->value;
 }
